requirements.txt
selenium==4.24.0
requests
pytest
allure-pytest